Create Survival_of_the_Biggest_number
The program receives, on the first line list of integer numbers (separated by a single space). On the second line, number n. The number "n" represents the count of numbers to remove from the list. The goal is to remove the smallest ones of the given numbers. Finally, print all remaining numbers in the list, separated by a comma and a space ", ".