Create Josephus_Permutation
People are standing in a circle. Counting begins from the first and proceeds from left to right throughout specific steps. After skipping an exact number of people, the "n" person is executed. The procedure is repeated with the remaining people, starting with the next person, going in the same direction, and skipping the same number of people until no one remains. The printed result is the order of executions.