Social Distribution - task
The first input is the population (numbers separated by comma and space ", "). 
The second input is the minimum wealth. 
The aim is to distribute the wealth so that no part of the population has less than the minimum wealth. 
For this purpose, always take wealth from the wealthiest part of the population. 
There will be cases where the distribution will not be possible. In that case, the print result is: "No equal distribution possible".